feat(pipeline): wire evaluate→session→polish/export end-to-end, add score .int() (B5)
把 B0~B4 引擎接成端到端 pipeline，修复四处断链：evaluate 段产出 EvaluationSummary
写入 session，驱动 polish 排序与 export 单页排版。summary 缺省全程优雅降级。

- coach/evaluation-research.ts（新）：抽 runEvaluationResearch + buildResearchFindings
  共享 helper（从 action-helpers.ts 内联三路计算 + findings 映射平移，逐字节等价）
- coach/action-helpers.ts：executeSearchEvaluation 改组合调用（对外契约零变化）；
  新增 executePipelineEvaluation——复用 helper 单次搜索 → synthesizeEvaluationSummary
  （config=null 规则降级）→ updateSessionEvaluationSummary 写 session.evaluationSummary
- pipeline/actions.ts：evaluate 切到 executePipelineEvaluation；polish 传
  { evaluationSummary }；export 传 { evaluationSummary, singlePage: true }
- pipeline/types.ts：experienceValueRatingSchema.score + jdMatchScore 加 .int()
  （引擎全过 clampScore=Math.round，安全；B2 遗留收口）

验收：scripts/orchestration-check.ts 18/18（闭网，config=null 规则降级，不出网）；
B0~B4 回归全过（.int() 后 evaluation-check 46/46、contracts-check 19/19 重点验）；
tsc/eslint/build 全绿。typescript-reviewer 1 MEDIUM（synthesizeEvaluationSummary
未包 try/catch 破坏 Promise<never> 契约）+ 3 LOW 均已修复。

一致性隐患（两套推进路径 completedAt 终态、session 唯一性、egress 两处同步、
throttleMs 未接入）记入 design §7 后续小批次，不在 B5 范围。

+++ app/src/features/pipeline/types.ts
@@ -90,7 +90,7 @@ export type AutoAdvanceConfig = z.infer<typeof autoAdvanceConfigSchema>;
 
 export const experienceValueRatingSchema = z.object({
   experienceId: z.string(),
-  score: z.number().min(0).max(100),
+  score: z.number().int().min(0).max(100),
   tier: z.enum(["high", "medium", "low"]),
   rationale: z.string(),
   searchCitations: z.array(z.string()),
@@ -103,7 +103,7 @@ export const evaluationSummarySchema = z.object({
   reportId: z.string(),
   createdAt: z.string(),
   experienceRatings: z.array(experienceValueRatingSchema),
-  jdMatchScore: z.number().min(0).max(100).optional(),
+  jdMatchScore: z.number().int().min(0).max(100).optional(),
   uncoveredKeywords: z.array(z.string()),
 });
